classmanager/requirements.txt
beautifulsoup4==4.9.1
cffi==1.14.0
Django==2.2.4
django-bootstrap4==2.1.1
django-misaka==0.2.1
djangorestframework==3.11.0
houdini.py==0.1.0
importlib-metadata==1.7.0
misaka==2.1.1
Pillow==7.1.2
pycparser==2.20
Pygments==2.6.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0

